+++ src/DataTable/SelectionStatus.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 import { Button } from '..';
-import DataTableContext from './TableContext';
+import DataTableContext from './DataTableContext';
 
 export const SELECT_ALL_TEST_ID = 'test_selection_state_select_all_button';
 export const CLEAR_SELECTION_TEST_ID = 'test_selection_state_clear_selection_button';
